\name{documented-package}
\alias{documented-package}
\alias{documented}
\docType{package}
\title{
\packageTitle{documented}
}
\description{
\packageDescription{documented}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{documented}
\packageIndices{documented}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{documented}

Maintainer: \packageMaintainer{documented}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
